sky_axs_initial_latest/core/ai_ops/metrics_collector/requirements.txt
redis==5.0.1
rq==1.16.2
requests==2.31.0
python-dateutil==2.9.0.post0
